Enhance API account balance endpoint
Why:

* Currently, the API account balance endpoint allows users to get the
ether balance for a single address. This commit introduces support for
getting the ether balance for multiple addresses (maximum of 20) in a
single batch.
* Issue link: https://github.com/poanetwork/poa-explorer/issues/138

This change addresses the need by:

* Adding `Chain.hashes_to_addresses/1` to get multiple addresses from
the DB, given a list of hashes.
* Editing `balance/2` in `API.RPC.AddressController` to support multiple
addresses. Previously it only supported a single address.
* Editing `API.RPC.AddressView` to handle a single address differently
than multiple addresses. This is to mimic Etherscan's API.