Edit API account balance actions to return Wei
Why:

* Currently, the `account#balance` and `account#balancemulti` API calls
return the Ether balance. It was built this way initially because the
Etherscan docs say those API calls should return Ether. This seems to be
a bug in their documentation because the API responds with the balance
in Wei.
* Issue link: related to
https://github.com/poanetwork/poa-explorer/issues/138

This change addresses the need by:

* Editing the `account#balance` and `account#balancemulti` API calls to
return Wei instead of Ether.
* Editing `API.RPC.AddressControllerTest` to assert account balances are
returned in Wei.