Remove duplicated code from Transaction page
The code that defines the card header, where the tab links are, was
being repeated in every tab. This commit moves this code to a partial
and creates a little helper function to get the currently active tab. This
function simply looks for the current tab name in the request path.