Ganache ignores internal transactions
Why:

* We were using the `EthereumJSONRPC.Geth` variant for Ganache but we
recently learned that it doesn't support the `tracer` option for
[debug_traceTransaction](https://github.com/ethereum/go-ethereum/wiki/Management-APIs#debug_tracetransaction).
This `tracer` option is what we currently use to fetch internal
transactions for Geth. This was the cause of errors as shown on issue
1152 (link below).
* Issue link:
https://github.com/poanetwork/blockscout/issues/1152

This change addresses the need by:

* Creating `EthereumJSONRPC.Ganache`. It ignores fetching beneficiaries,
internal transactions, and pending transactions.
* Editing ganache config to use the `EthereumJSONRPC.Ganache` variant
instead of `EthereumJSONRPC.Geth`.
* Updating inaccurate doc in `EthereumJSONRPC.Geth`. It previously said
fetching internal transactions was not supported for Geth which is
inaccurate.