Set Token holder_count to 0 for new tokens
Fixes #1359

If there is a conflict, then the token is no new and existing
holder_count is not overriden.  This ensures that pre-existing tokens
that haven't been data migrated are kept as `holder_count` `NULL` and
migrated tokens or new tokens passing through this change leave the
`holder_count` unchanged.